+++ mtg/src/main/scala/com/github/gchudnov/mtg/internal/RelStatic.scala
@@ -2,6 +2,7 @@ package com.github.gchudnov.mtg.internal
 
 import com.github.gchudnov.mtg.Domain
 import com.github.gchudnov.mtg.Interval
+import com.github.gchudnov.mtg.internal.rel.*
 
 /**
  * Static Interval Relations
@@ -12,4 +13,10 @@ private[mtg] transparent trait RelStatic:
    * Relation: Before, Precedes (b)
    */
   final def before[T: Domain](a: Interval[T], b: Interval[T]): Boolean =
-    a.before(b)
+    Before.before(a, b)
+
+  /**
+   * Relation: After, IsPrecededBy (B)
+   */
+  final def after[T: Domain](a: Interval[T], b: Interval[T]): Boolean =
+    Before.after(a, b)
