+++ mtg/src/main/scala/com/github/gchudnov/mtg/internal/rel/Before.scala
@@ -1,0 +1,49 @@
+package com.github.gchudnov.mtg.internal.rel
+
+import com.github.gchudnov.mtg.Interval
+import com.github.gchudnov.mtg.Domain
+
+private[mtg] object Before:
+
+  /**
+   * Before, Precedes (b)
+   *
+   * {{{
+   *   PP (Point-Point):
+   *   {p}; {q}
+   *   p < q
+   *
+   *   PI (Point-Interval):
+   *   {p}; {a-, a+}
+   *   p < a-
+   *
+   *   II (Interval-Interval):
+   *   {a-, a+}; {b-, b+}
+   *   a- < b-
+   *   a- < b+
+   *   a+ < b-
+   *   a+ < b+
+   *
+   *   a- < a+ < b- < b+
+   *
+   *   Relation                  AAAAA
+   *   before(a,b)      b        :   : BBBBBBBBB  |  a+ < b-
+   * }}}
+   * 
+   * @param a interval
+   * @param b interval
+   * @return true if a is before b
+   */
+  final def before[T: Domain](a: Interval[T], b: Interval[T]): Boolean =
+    val ordM = summon[Domain[T]].ordMark
+    a.nonEmpty && b.nonEmpty && ordM.lt(a.right, b.left)
+
+  /**
+    * After, Follows (a)
+    *
+    * @param a interval
+    * @param b interval
+    * @return true if a is after b
+    */
+  final def after[T: Domain](a: Interval[T], b: Interval[T]): Boolean =
+    before(b, a)
