+++ mtg/src/test/scala/com/github/gchudnov/mtg/internal/IsLessSpec.scala
@@ -5,6 +5,7 @@ import com.github.gchudnov.mtg.Interval
 import com.github.gchudnov.mtg.Domain
 import com.github.gchudnov.mtg.TestSpec
 import org.scalatestplus.scalacheck.ScalaCheckPropertyChecks.*
+import com.github.gchudnov.mtg.internal.rel.IntervalRelAssert
 
 /**
  * IsLess
